Improve admin in react, better separation of concerns and all lists working
