+++ admin/src/config/apiPoints.js
@@ -1,0 +1,3 @@
+
+export const entryPoint = 'http://localhost:8096/api';
+export const loginPoint = 'http://localhost:8096/api/login_check';
